Added a very basic example of Factory Pattern
This very basic example of a factory pattern in action will demostrate
how programming to the interface instead of the concrete implementation
means we are not bound to either the Dog or Cat type. As long as both
classes implement the same interface, mypet can be either a Dog or a
Cat.